backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hub-backend"
version = "0.1.0"
description = "HUB 数据中台后端"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.110,<0.120",
    "uvicorn[standard]>=0.27",
    "tortoise-orm[asyncpg]>=0.20,<0.22",
    "aerich>=0.7,<0.8",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-multipart>=0.0.6",
    "httpx>=0.26",
    "redis>=5.0",
    "cryptography>=42.0",
    "passlib[bcrypt]>=1.7",
    "python-dateutil>=2.8",
    "structlog>=24.1",
    "dingtalk-stream>=0.18",
    "tiktoken>=0.5",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    # Plan 6 ToolRegistry 用 jsonschema 验 tool args（registry.py:_validate_args）
    # 本地 venv 因 transitive 装了所以一直可用；docker 镜像不带，必须显式声明
    "jsonschema>=4.0",
    # Plan 6 eval runner 加载 gold_set.yaml；当前由 dingtalk-stream / aerich 间接拉到，
    # 但这是 Plan 6 自身用的依赖——显式声明防上游断链
    "PyYAML>=6.0",
    # Plan 6 v9 GraphAgent：LangGraph 状态机 + LangChain Core 基础类型
    "langgraph>=0.2.0,<0.3",
    "langchain-core>=0.3.0,<0.4",
    "fakeredis>=2.35.1",
    # Plan 6 v9 跨 worker 重启 checkpoint 持久化（之前用 in-process MemorySaver 重启即丢）
    "langgraph-checkpoint-postgres>=2.0.0,<3",
    "psycopg[binary]>=3.1",
    "psycopg_pool>=3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "fakeredis>=2.20",
    "ruff>=0.3",
    "mypy>=1.8",
]

# setuptools 包发现：仅扫描 hub/ 这个 namespace，避免把 tests/ migrations/ 当包安装
[tool.setuptools.packages.find]
where = ["."]
include = ["hub", "hub.*"]
exclude = ["tests*", "migrations*"]

[tool.aerich]
tortoise_orm = "hub.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "eval: 真实 LLM eval 测试（CI 默认跳过；Task 19 单独跑 pytest -m eval）",
    "realllm: 真实 LLM 集成测试（需要 DEEPSEEK_API_KEY；无 key 时全 skip）",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
